Rename all commands to their full class names
