+++ source/classes/commands/CommentCommand.js
