+++ source/classes/commands/CreateEntryCommand.js
