+++ source/classes/commands/DeleteEntryAttributeCommand.js
