+++ source/classes/commands/DeleteEntryCommand.js
