+++ source/classes/commands/DeleteEntryMetaCommand.js
