+++ source/classes/commands/DeleteGroupAttributeCommand.js
