+++ source/classes/commands/DeleteGroupCommand.js
