+++ source/classes/commands/FormatCommand.js
