+++ source/classes/commands/MoveEntryCommand.js
