+++ source/classes/commands/MoveGroupCommand.js
