+++ source/classes/commands/SetEntryAttributeCommand.js
