+++ source/classes/commands/SetEntryMetaCommand.js
