+++ source/classes/commands/SetEntryPropertyCommand.js
