+++ source/classes/commands/SetGroupAttributeCommand.js
