+++ source/classes/commands/TitleGroupCommand.js
